Fixed issue with Eloquent::increment and decrement
When calling `increment('column')` on an Eloquent model instance, I expect only the column of the row of that model to be incremented. This however is not the case, you'd have to explicitly set the where clause yourself before calling `increment`.

When `increment` is called on the Eloquent model, it is magically routed to `Eloquent\Query` from where it is directly routed to `Query`. From there the table is immediately updated and the result is returned. This commit sets the where clause in `Eloquent\Query` just before the operation is routed to `Query` and executed, thus resolving the problem.